multi-store feature

+++ calculate.py
@@ -40,7 +40,7 @@ def calculate(course, quiz, lab, assignment, presentation, participation, midter
             
 
     if not os.path.exists("./users/{0}.json".format(user_name)):
-        create_json = {
+        create_json = [{
                             "course_name": "",
                             "quiz": 0,
                             "lab": 0,
@@ -51,40 +51,65 @@ def calculate(course, quiz, lab, assignment, presentation, participation, midter
                             "final": 0,
                             "total": 0,
                             "date": ""
-                        }
+                        }]
         with open("./users/{0}.json".format(user_name), 'w') as f:
             json.dump(create_json, f, indent=4)
         with open("./users/{0}.json".format(user_name), 'r+') as f:
             file_data = json.load(f)
-            file_data['course_name'] = course
-            file_data['quiz'] = quiz_mark
-            file_data['lab'] = lab_mark
-            file_data['assignments_projects'] = assignment_mark
-            file_data['presentations'] = presentation_mark
-            file_data['participation'] = participation_mark
-            file_data['midterm'] = midterm_mark
-            file_data['final'] = final_mark
-            file_data['total'] = total_mark
-            file_data['date'] = today.strftime("%m/%d/%y")
-            f.seek(0)
-            json.dump(file_data, f, indent =4)
-            f.truncate()
+            for i in file_data:
+                i['course_name'] = course
+                i['quiz'] = quiz_mark
+                i['lab'] = lab_mark
+                i['assignments_projects'] = assignment_mark
+                i['presentations'] = presentation_mark
+                i['participation'] = participation_mark
+                i['midterm'] = midterm_mark
+                i['final'] = final_mark
+                i['total'] = total_mark
+                i['date'] = today.strftime("%m/%d/%y")
+                f.seek(0)
+                json.dump(file_data, f, indent =4)
+                f.truncate()
     else:
         with open("./users/{0}.json".format(user_name), 'r+') as f:
             file_data = json.load(f)
-            file_data['course_name'] = course
-            file_data['quiz'] = quiz_mark
-            file_data['lab'] = lab_mark
-            file_data['assignments_projects'] = assignment_mark
-            file_data['presentations'] = presentation_mark
-            file_data['participation'] = participation_mark
-            file_data['midterm'] = midterm_mark
-            file_data['final'] = final_mark
-            file_data['total'] = total_mark
-            file_data['date'] = today.strftime("%m/%d/%y")
-            f.seek(0)
-            json.dump(file_data, f, indent =4)
-            f.truncate()
+            user_list = []
+            for i in file_data:
+                user_list.append(i['course_name'])
+            if course not in user_list:
+                to_add = {
+                            "course_name": course,
+                            "quiz": quiz_mark,
+                            "lab": lab_mark,
+                            "assignments_projects": assignment_mark,
+                            "presentations": presentation_mark,
+                            "participation": participation_mark,
+                            "midterm": midterm_mark,
+                            "final": final_mark,
+                            "total": total_mark,
+                            "date": today.strftime("%m/%d/%y")
+                }
+                file_data.append(to_add)
+                f.seek(0)
+                json.dump(file_data, f, indent =4)
+                f.truncate()
+            else:
+                for i in file_data:
+                    if i['course_name'] == course:
+                        i['quiz'] = quiz_mark
+                        i['lab'] = lab_mark
+                        i['assignments_projects'] = assignment_mark
+                        i['presentations'] = presentation_mark
+                        i['participation'] = participation_mark
+                        i['midterm'] = midterm_mark
+                        i['final'] = final_mark
+                        i['total'] = total_mark
+                        i['date'] = today.strftime("%m/%d/%y")
+                        f.seek(0)
+                        json.dump(file_data, f, indent =4)
+                        f.truncate()
+
+                        
     return
 
-calculate('ACIT 1420',100,20,30,50,10,20,10)
+calculate('ACIT 1515',50,20,30,50,10,30,10)
